lesson_6 methods for TicTacToe moved to additional class AI
